EliteMobs 8.6.1-SNAPSHOT-2 - [New] Based on feedback, gold shotguns no longer move around when triggered, they will aim in a specific direction and shoot after the delay similar to other powers, making it possible to dodge them - [New] Added individual damage multipliers for combat against bosses in the normalized combat system (regional bosses). This means that the damage multipliers for elites / event bosses and for dungeon content can and must now be adjusted individually. This update resets configuration changes to combat multipliers to avoid causing balance issues due to the new combat balance. - [Fix] Changed default flag setup to allow the use of redstone in regions set up by EliteMobs - [Tweak] Tweaked the mob level to use integer values for the level difference warning - [Fix] Fixed issue where events would no longer trigger at all - [Tweak] When regional bosses engage in combat with other bosses they will now lose the out of combat slowness effect - [Tweak] Gold shotguns no longer deal damage to other elites

+++ src/main/java/com/magmaguy/elitemobs/items/ItemLootShower.java
@@ -64,7 +64,7 @@ public class ItemLootShower implements Listener {
                     player.spigot().sendMessage(ChatMessageType.ACTION_BAR, TextComponent.fromLegacyText(
                             ChatColorConverter.convert(ItemSettingsConfig.getLevelRangeTooDifferent())
                                     .replace("$playerLevel", ElitePlayerInventory.playerInventories.get(player.getUniqueId()).getFullPlayerTier(false) + "")
-                                    .replace("$bossLevel", mobLevel + "")));
+                                    .replace("$bossLevel", (int) mobLevel + "")));
                 }
             }.runTaskTimer(MetadataHandler.PLUGIN, 0, 1);
             return;
